+++ adapters/dbmodel_to_apiv1.js
